fix(sudo): expose env to command being run instead of to sudo itself
Previously the way env was being handled, it was simply passed down to
the wrapped Runner instance. This means that env was set on the in the
context of sudo, rather than the command that sudo was running.

We now instead pass the set env as arguments to sudo, which exposes
those environment variables to the command it is running.