feat(ssh-cli): add ssh CLI based wrapper Runner implementation
SSHCLI is a Runner that wraps another Runner, essentially prefixing
given commands and arguments with "ssh", relevant SSH CLI arguments, and
the given destination. It then passes this new "ssh" command to the
underlying Runner.

This is useful for running commands on remote hosts via SSH, without
having to use the Go ssh package.

Interactive commands are not supported, meaning SSH password prompts
will not work, and the remote machine's hostkey should already be known
and trusted by the ssh CLI client.